chore: update file naming to the standard
