+++ internal/adminpanel/handlers/deleteCorruptedNzb.go
@@ -3,22 +3,22 @@ package handlers
 import (
 	"net/http"
 
-	corruptednzbsmanager "github.com/javi11/usenet-drive/internal/usenet/corrupted-nzbs-manager"
+	"github.com/javi11/usenet-drive/internal/usenet/corruptednzbsmanager"
 	echo "github.com/labstack/echo/v4"
 )
 
-type discardCorruptedNzbRequest struct {
+type deleteCorruptedNzbRequest struct {
 	Path string `json:"path"`
 }
 
-func DiscardCorruptedNzbHandler(cNzb corruptednzbsmanager.CorruptedNzbsManager) echo.HandlerFunc {
+func DeleteCorruptedNzbHandler(cNzb corruptednzbsmanager.CorruptedNzbsManager) echo.HandlerFunc {
 	return func(c echo.Context) error {
-		req := new(discardCorruptedNzbRequest)
+		req := new(deleteCorruptedNzbRequest)
 		if err := c.Bind(req); err != nil {
 			return echo.NewHTTPError(http.StatusBadRequest, err.Error())
 		}
 
-		if err := cNzb.Discard(c.Request().Context(), req.Path); err != nil {
+		if err := cNzb.Delete(c.Request().Context(), req.Path); err != nil {
 			return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
 		}
