+++ internal/usenet/connectionpool/config.go
