+++ internal/usenet/connectionpool/connectionpool.go
