+++ internal/usenet/filereader/errors.go
@@ -1,4 +1,4 @@
-package usenetfilereader
+package filereader
 
 import "errors"
