+++ internal/usenet/filereader/fileinfo.go
@@ -1,4 +1,4 @@
-package usenetfilereader
+package filereader
 
 import (
 	"fmt"
