+++ internal/usenet/filereader/filereader.go
@@ -1,12 +1,12 @@
-package usenetfilereader
+package filereader
 
 import (
 	"context"
 	"io/fs"
 	"log/slog"
 
-	connectionpool "github.com/javi11/usenet-drive/internal/usenet/connection-pool"
-	corruptednzbsmanager "github.com/javi11/usenet-drive/internal/usenet/corrupted-nzbs-manager"
+	"github.com/javi11/usenet-drive/internal/usenet/connectionpool"
+	"github.com/javi11/usenet-drive/internal/usenet/corruptednzbsmanager"
 	"github.com/javi11/usenet-drive/internal/usenet/nzbloader"
 	"golang.org/x/net/webdav"
 )
