+++ internal/usenet/filereader/utils.go
@@ -1,4 +1,4 @@
-package usenetfilereader
+package filereader
 
 import (
 	"os"
