+++ internal/usenet/filewriter/article.go
@@ -1,4 +1,4 @@
-package usenetfilewriter
+package filewriter
 
 import (
 	"bytes"
